pyproject.toml
[project]
name = "kinspot-edgeai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.6"
dependencies = [
    "accelerate>=1.10.1",
    "deepface>=0.0.99",
    "faiss-cpu>=1.13.0",
    "matplotlib>=3.9.4",
    "mlxtend>=0.23.4",
    "opencv-python>=4.13.0.92",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pillow-heif>=1.1.1",
    "scikit-learn>=1.6.1",
    "setuptools<82",
    "tf-keras>=2.20.1",
    "torch>=2.8.0",
    "torchmetrics>=1.8.2",
    "torchvision>=0.23.0",
    "transformers>=4.57.6",
]
